test: share begin_query_run and query_completed_action fixtures in query mod
